move ScheduleProvider into services.scheduleService
